aco: use size_t for monotonic_buffer_resource
Necessary for really big shaders.

Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/14650
Backport-to: 25.3
(cherry picked from commit 874255e89964c8436789d153df2432127d5c13af)

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/39462>

+++ src/amd/compiler/aco_util.h
@@ -274,7 +274,7 @@ public:
       }
 
       /* create new larger buffer */
-      uint32_t total_size = buffer->data_size + sizeof(Buffer);
+      size_t total_size = buffer->data_size + sizeof(Buffer);
       do {
          total_size *= 2;
       } while (total_size - sizeof(Buffer) < size);
@@ -302,8 +302,8 @@ public:
 private:
    struct Buffer {
       Buffer* next;
-      uint32_t current_idx;
-      uint32_t data_size;
+      size_t current_idx;
+      size_t data_size;
       uint8_t data[];
    };
 
@@ -333,7 +333,7 @@ public:
    /* Memory Allocation */
    T* allocate(size_t size)
    {
-      uint32_t bytes = sizeof(T) * size;
+      size_t bytes = sizeof(T) * size;
       return (T*)memory_resource.get().allocate(bytes, alignof(T));
    }
